requirements.txt
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0
google-api-python-client
